+++ test/testPackageFiles.js
